Refactor: Add WorkspaceImageInput component
