+++ src/components/admin/workspace/WorkspaceEdit.tsx
@@ -10,9 +10,8 @@ import { adminWorkspaceAtom } from '@recoils/atoms';
 import { Color } from '@resources/colors';
 import { colFlex, rowFlex } from '@styles/flexStyles';
 import { WorkspaceImage } from '@@types/index';
-import PlusIconSvg from '@resources/svg/PlusIconSvg';
-import { expandButtonStyle } from '@styles/buttonStyles';
 import { removeAndPushNull } from '@utils/workspaceEdit';
+import WorkspaceImageInput from './WorkspaceImageInput';
 
 const ContentContainer = styled.div`
   width: 100%;
@@ -65,34 +64,6 @@ const ImageInputContainer = styled.div`
   width: 85%;
   height: 100%;
   ${rowFlex({ justify: 'space-between', align: 'start' })}
-`;
-
-const ImageInput = styled.img`
-  width: 260px;
-  height: 100%;
-  border-radius: 10px;
-  border: none;
-  background: ${Color.LIGHT_GREY};
-  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.1) inset;
-  cursor: pointer;
-  ${rowFlex({ justify: 'center', align: 'center' })}
-`;
-
-const ImageDummyInput = styled.div`
-  width: 260px;
-  height: 100%;
-  border-radius: 10px;
-  border: none;
-  background: ${Color.LIGHT_GREY};
-  box-shadow: 0px 0px 5px 0px rgba(0, 0, 0, 0.1) inset;
-  ${rowFlex({ justify: 'center', align: 'center' })}
-  cursor: pointer;
-`;
-
-const PlusIcon = styled(PlusIconSvg)`
-  width: 30px;
-  height: 30px;
-  ${expandButtonStyle};
 `;
 
 const DescriptionContainer = styled.div`
@@ -289,15 +260,7 @@ function WorkspaceEdit() {
             <AppLabel size={20}>대표 사진</AppLabel>
           </ImageLabelContainer>
           <ImageInputContainer>
-            {displayImages.map((img, index) =>
-              img ? (
-                <ImageInput key={img.id} src={img.url} onClick={() => handleImageClick(index)} />
-              ) : (
-                <ImageDummyInput key={`${index}-dummy`} onClick={() => handleImageClick(index)}>
-                  <PlusIcon />
-                </ImageDummyInput>
-              ),
-            )}
+            <WorkspaceImageInput images={displayImages} handleImageClick={handleImageClick} />
           </ImageInputContainer>
         </ImageContainer>
         <DescriptionContainer>
